sm8450-common: Specify an up to date architecture
The Cortex-A510 and other cores of the Snapdragon 8 Gen 1 cause Clang
miscompilations, but specifying a Cortex-A76 should be an improvement
already. Additionally, we have an ARMv9-a CPU, so target that; even if
AOSP hacks it as equivalent to ARMv8.2-a+dotprod with branch protection,
it is already an improvement.

Change-Id: I1d1d69d57a31ef163c681918c15fcf19e1708358
Signed-off-by: Jens Reidel <adrian@travitia.xyz>